requirements.txt
absl-py==2.0.0
accelerate==0.26.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apturl==0.5.2
asttokens==2.4.1
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.13
botocore==1.34.13
Brlapi==0.8.3
bs4==0.0.2
cached-property==1.5.2
cachetools==5.3.2
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
comm==0.2.1
command-not-found==0.3
commonmark==0.9.1
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
deta==1.2.0
dictdiffer==0.9.0
diffusers==0.25.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.5.0
duckdb==0.9.2
duplicity==0.8.21
email-validator==2.1.0.post1
exceptiongroup==1.2.0
executing==2.0.1
extra-streamlit-components==0.1.60
fasteners==0.14.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
fonttools==4.47.0
fsspec==2023.12.2
future==0.18.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.26.1
google-auth-oauthlib==1.2.0
gpg==1.16.0
grpcio==1.60.0
gspread==5.12.4
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.2.3
gyp==0.1
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httpx==0.26.0
huggingface==0.0.1
huggingface-hub==0.20.2
idna==2.10
imbalanced-learn==0.12.0
imblearn==0.0
importlib-metadata==4.6.4
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpickle==3.0.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
music21==9.1.0
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
notion==0.0.28
notion-client==2.2.1
notion-sdk==0.7.0
notion-sdk-py==0.0.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
packaging==21.3
pandas==1.5.3
paramiko==2.9.3
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
plotly-express==0.4.1
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybboxes==0.1.6
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-magic==0.4.27
python-slugify==8.0.3
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.32.0
regex==2023.12.25
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
roboflow==1.1.15
rpds-py==0.16.2
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
sahi==0.11.15
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
SecretStorage==3.3.1
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sql_metadata==2.10.0
sqlparse==0.4.4
st-gsheets-connection==0.0.3
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.30.0
streamlit-authenticator==0.2.3
streamlit-autorefresh==1.0.1
supervision==0.17.1
sympy==1.12
systemd-python==234
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torbrowser-launcher==0.3.3
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ultralytics==8.0.235
unattended-upgrades==0.1
urllib3==1.26.4
usb-creator==0.3.7
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.12
webcolors==1.13
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
yolov5==7.0.13
zipp==1.0.0
